+++ changelog.md
@@ -6,6 +6,10 @@
 
 - Use encryption for the globale Dropbox token
 
+### Fixed
+
+- Fixed missing permission check of the abilities (thanks to Philipp Beick)
+
 ## [5.2.2] - 27.06.2026
 
 ### Changed